move suggestion list tests
