+++ src/app/views/query-runner/query-input/auto-complete/suggestion-list/SuggestionsList.tsx
@@ -1,9 +1,9 @@
 import { Label, styled } from '@fluentui/react';
 import React, { useEffect } from 'react';
 
-import { ISuggestionsList } from '../../../../../types/auto-complete';
-import { classNames } from '../../../classnames';
-import { autoCompleteStyles } from './auto-complete.styles';
+import { ISuggestionsList } from '../../../../../../types/auto-complete';
+import { classNames } from '../../../../classnames';
+import { autoCompleteStyles } from '../auto-complete.styles';
 
 const SuggestionsList = (props: any) => {
   const { filteredSuggestions, activeSuggestion, onClick }: ISuggestionsList = props;
